refactor: use express router for routing purposes
